Display scale fix
Some parts on the form do not show up or look messed up when the display scale is not the default 100%.
Moved the log at the bottom to a sub-form, but this only made the issue worse.
Setting the AutoScaleMode on the main form from Font to DPI and to Inherit on the sub forms seems to finally fix this.